merge: 合并master存档记录ID溢出修复(6ca6c4011)
master改动:
- 存档相关ID由int改long: inventoryitems/inventoryequipment/inventorymerchant/
  dueyitems/keymap/skills/buddies/skillmacros/savedlocations等
- 9个DO实体id字段改Long
- 新增幂等BIGINT迁移SQL

冲突解决:
- ItemFactory.java: 保留HEAD的rs.next()防御检查,同时应用master的setLong;
  冲突块后紧跟的psEquip.setInt(1,rs.getInt(1))一并改为setLong,否则BIGINT
  修复会被覆盖

重要修正 - 迁移版本号冲突:
- 仓库已存在 V1.11.7__fix_alliance_membership_transactions.sql 与
  V1.11.8__add_character_load_lookup_indexes.sql(见b5b40d75b曾修同类问题)
- master新增脚本撞同版本号,Flyway会因同名多版本错误导致启动失败
- 顺延为 V1.11.9__queststatusid_bigint.sql 与
  V1.11.10__save_record_ids_bigint.sql

+++ gms-server/src/main/java/org/gms/client/inventory/ItemFactory.java
@@ -263,7 +263,7 @@ public enum ItemFactory {
 
                                 if (mit.equals(InventoryType.EQUIP) || mit.equals(InventoryType.EQUIPPED)) {
                                     Equip equip = (Equip) item;
-                                    psEquip.setInt(1, rs.getInt(1));
+                                    psEquip.setLong(1, rs.getLong(1));
                                     psEquip.setInt(2, equip.getUpgradeSlots());
                                     psEquip.setInt(3, equip.getLevel());
                                     psEquip.setInt(4, equip.getStr());
@@ -339,7 +339,7 @@ public enum ItemFactory {
                     while (rs.next()) {
                         short bundles = 0;
                         try (PreparedStatement psBundle = con.prepareStatement("SELECT `bundles` FROM `inventorymerchant` WHERE `inventoryitemid` = ?")) {
-                            psBundle.setInt(1, rs.getInt("inventoryitemid"));
+                            psBundle.setLong(1, rs.getLong("inventoryitemid"));
 
                             try (ResultSet rs2 = psBundle.executeQuery()) {
                                 if (rs2.next()) {
@@ -408,7 +408,7 @@ public enum ItemFactory {
                     final InventoryType mit = pair.getRight();
                     i++;
 
-                    final int genKey;
+                    final long genKey;
                     // Item
                     try (PreparedStatement ps = con.prepareStatement("INSERT INTO `inventoryitems` VALUES (DEFAULT, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)", Statement.RETURN_GENERATED_KEYS)) {
                         ps.setInt(1, value);
@@ -430,13 +430,13 @@ public enum ItemFactory {
                                 throw new RuntimeException("Inserting item failed.");
                             }
 
-                            genKey = rs.getInt(1);
+                            genKey = rs.getLong(1);
                         }
                     }
 
                     // Merchant
                     try (PreparedStatement ps = con.prepareStatement("INSERT INTO `inventorymerchant` VALUES (DEFAULT, ?, ?, ?)", Statement.RETURN_GENERATED_KEYS)) {
-                        ps.setInt(1, genKey);
+                        ps.setLong(1, genKey);
                         ps.setInt(2, id);
                         ps.setInt(3, bundles);
                         ps.executeUpdate();
@@ -445,7 +445,7 @@ public enum ItemFactory {
                     // Equipment
                     if (mit.equals(InventoryType.EQUIP) || mit.equals(InventoryType.EQUIPPED)) {
                         try (PreparedStatement ps = con.prepareStatement("INSERT INTO `inventoryequipment` VALUES (DEFAULT, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)")) {
-                            ps.setInt(1, genKey);
+                            ps.setLong(1, genKey);
 
                             Equip equip = (Equip) item;
                             ps.setInt(2, equip.getUpgradeSlots());
